set peer ip address
For p2p links with each endpoint in a different IP subnet, as would be
found with IPv4 addresses and 32-bit subnet masks, configure the peer
address.  In Linux, adding an interface route with the peer's address
as the destination is not sufficient; no other route will be able to
use the peer's address as a next-hop.