munet: improve sshremote node
Previously only supported reaching nodes within the topology which
was actually a very uncommon use case (but the one that we can easily
test in CI).

Now we support reaching the server using the host network namespace.
It is far more likely that a router or server will be reachable over
a mgmt network connected to the host than within the munet network
namespace.