mutini: handle possible missed zombie cleanup
Rather than reaping on SIGCHLD, use a poll() with a timeout that will
check for zombies, the poll() exits on SIGCHLD and also on timeout. This
seems to catch the case we had before were we missed cleaning up a
zombie

Also we avoid logging in the signal handler which was causing an
exception.

Signed-off-by: Christian Hopps <chopps@labn.net>