munet: make Commander.spawn() async
spawn(), which conducts a send/expect process is
blocking due to the abscence of async_=True to
pexpect.expect() within the method's main
send/expect loop. This is undesirable due the
presence of other ongoing initialization
coroutines, since it will block their progress
for the entire time needed to complete either
the single nodes send/expect loop (or wait for
it to timeout).

This is particularly problematic in cases where
multiple VMs requiring custom console expect/send
prompts are required (e.g. routers). Since each
QEMU VM sets up both console logging and completes
an expect/send loop without yielding control to
the coroutine of the other QEMU VMs, only one VM
can ever have logging configured (and thus advance
in an expect/send loop) at any given time. Not
only is this inefficient given that all QEMU VMs
are running and waiting for input, but this can
be fatal if console logging is not set up on each
VM in time. If some mandatory console output is
missed, then there is no way to recover and the
setup of the QEMU VM will time out.

By changing spawn() such that is awaits for
pexpect.expect(), control can be properly passed
to coroutines of other QEMU VM nodes. This allows
for logging to be set up on all nodes immediately
and for independent progress to be made in each
VM's expect/send loop.

Signed-off-by: Liam Brady <lbrady@labn.net>