docs: Alternate install instructions
When updates are made on master but no release
cut, it is unclear to a user as to how to
aquire the recent fixes or changes. The set of
developer steps is sufficient to run the unit
tests, but is not applicable to a wide range
of dev enviornments where munet/mutest is in
use.

Conveniently, you can install munet with all
the latest changes from github using pip.
This change guides the user to such an option
if they need the latest updates.

Signed-off-by: Liam Brady <lbrady@labn.net>