munet: fix desync after send/expect
spawn() within shell_spawn() already provides
a newline when no shell/login prompt is found.
Sending a second newline after the expected
prompt is detected results in a desync in the
console I/O. This is because the response
received from sending the newline is not
consumed.

Thus, any future instance after the desync
of cmd_nostatus() or cmd_status() after
shell_spawn() finishes either throws a
warning or fails. The instability due to
the desync can break following features that
rely on such functionality, such as the
mounting of shared volumes into the QEMU VM.

Signed-off-by: Liam Brady <lbrady@labn.net>