mutest: mutest groups
Multiple mutests found within a subdirectory of a munet topology are
newly interpreted as a single 'group' of mutests. A group of mutests
share the same munet topology instance when running. This allows a
developer to avoid the cost of spinning up a new topology for each
mutest where it is not needed.

Signed-off-by: Liam Brady <lbrady@labn.net>