munet: run mutests from the munet cli
Refactor mutest such that the 'execute_test' function can be called from
a non-mutest context. Specifcally, it is modified such that it can be
called from the munet CLI, allowing a mutest to be run repeatedly when
drafting the test. Alongside this, a new Completer is added to detect
local mutest files and auto-complete the command.

Signed-off-by: Liam Brady <lbrady@labn.net>