munet: simplify async child watcher setting for old pythons
Older pytest-asyncio did things with event_loops that we tried to work with, but
in the end this didn't work and pytest-asyncio was improved to not require it
anymore.

Now we just have to make sure we have a non-threaded child watcher for asyncio
when we are unsharing inline with a new pid namespace as we are not allowed to
create threads in the main process under these conditions.

Signed-off-by: Christian Hopps <chopps@labn.net>