munet: fix --pcaps=all
Fix a bug where self.switches.keys() does not have the split() method,
and as a result, throws an error when 'all' is passed instead of the
specific networks/interfaces.

Note that 'all' will include all networks, but does not start pcaps for
specific interfaces. Interfaces must still be specified as NODE:INTF.

Signed-off-by: Liam Brady <lbrady@labn.net>